Start resolving ambiguity in datatypes

+++ src/communication.hpp
@@ -9,8 +9,8 @@ void initComm();
 void sendData(std::array<uint8_t, 64> &tx_data);
 void packGameSample(const std::array<float, 15> &game_sample, std::array<uint8_t, 64> &tx_data);
 void sendSample(const std::array<float, 15> &game_sample, std::array<uint8_t, 64> &tx_data);
-void packRawSample(unsigned long timestamp, int deviation, const std::array<uint16_t, 20> &raw_sample, std::array<uint8_t, 64> &tx_data);
-void sendSample(unsigned long timestamp, int deviation, std::array<uint16_t, 20> &raw_sample, std::array<uint8_t, 64> &tx_data);
+void packRawSample(uint32_t timestamp, int16_t deviation, const std::array<uint16_t, 20> &raw_sample, std::array<uint8_t, 64> &tx_data);
+void sendSample(uint32_t timestamp, int16_t deviation, std::array<uint16_t, 20> &raw_sample, std::array<uint8_t, 64> &tx_data);
 int receiveData(std::array<uint8_t, 64> &rx_data);
 };
